% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcep-package.R
\docType{package}
\name{wcep-package}
\alias{wcep-package}
\title{wcep: Survival Analysis for Weighted Composite Endpoints}
\description{
Analyze given data frame with multiple endpoints and return Kaplan-Meier survival probabilities together with the specified confidence interval. See Nabipoor M, Westerhout CM, Rathwell S, and Bakal JA (2023) \doi{10.1186/s12874-023-01857-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sarah-0k/wcep}
  \item Report bugs at \url{https://github.com/sarah-0k/wcep/issues}
}

}
\author{
\strong{Maintainer}: Sarah Rathwell \email{srathwel@ualberta.ca} (\href{https://orcid.org/0000-0003-3176-994X}{ORCID})

Authors:
\itemize{
  \item Jeffrey Bakal \email{jbakal@ualberta.ca} (\href{https://orcid.org/0000-0002-3658-2554}{ORCID})
  \item Cynthia Westerhout \email{cindy.westerhout@ualberta.ca} (\href{https://orcid.org/0000-0003-1082-5981}{ORCID})
  \item Caroline Falvey \email{caroline.falvey@duke.edu}
  \item Huiman Barnhart \email{huiman.barnhart@duke.edu}
}

Other contributors:
\itemize{
  \item Na Zhang \email{Na.Zhang@healthsharedservices.ca} (\href{https://orcid.org/0000-0001-7850-280X}{ORCID}) [contributor]
}

}
\keyword{internal}
