% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcep-package.R
\docType{package}
\name{wcep-package}
\alias{wcep}
\alias{wcep-package}
\title{wcep: Survival Analysis for Weighted Composite Endpoints}
\description{
Analyze given data frame with multiple endpoints and return Kaplan-Meier survival probabilities together with the specified confidence interval.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sarah-0k/wcep}
  \item Report bugs at \url{https://github.com/sarah-0k/wcep}
}

}
\author{
\strong{Maintainer}: Sarah Rathwell \email{srathwel@ualberta.ca} (\href{https://orcid.org/0000-0003-3176-994X}{ORCID})

Authors:
\itemize{
  \item Majid Nabipoor \email{nabipoor@ualberta.ca} (\href{https://orcid.org/0000-0003-2578-8300}{ORCID})
  \item Cynthia Westerhout \email{cindy.westerhout@ualberta.ca} (\href{https://orcid.org/0000-0003-1082-5981}{ORCID})
  \item Jeffrey Bakal \email{jbakal@ualberta.ca} (\href{https://orcid.org/0000-0002-3658-2554}{ORCID})
}

}
\keyword{internal}
